+++ src/Configuration/Cache.php
@@ -23,7 +23,7 @@ namespace TechDivision\Import\Configuration\Jms\Configuration;
 use JMS\Serializer\Annotation\Type;
 use JMS\Serializer\Annotation\SerializedName;
 use TechDivision\Import\Utils\CacheTypes;
-use TechDivision\Import\ConfigurationInterface;
+use TechDivision\Import\Configuration\ConfigurationInterface;
 use TechDivision\Import\Configuration\CacheConfigurationInterface;
 
 /**
@@ -66,7 +66,7 @@ class Cache implements CacheConfigurationInterface
     /**
      * A reference to the parent configuration instance.
      *
-     * @var \TechDivision\Import\ConfigurationInterface
+     * @var \TechDivision\Import\Configuration\ConfigurationInterface
      */
     protected $configuration;
 
@@ -106,7 +106,7 @@ class Cache implements CacheConfigurationInterface
     /**
      * Set's the reference to the configuration instance.
      *
-     * @param \TechDivision\Import\ConfigurationInterface $configuration The configuration instance
+     * @param \TechDivision\Import\Configuration\ConfigurationInterface $configuration The configuration instance
      *
      * @return void
      */
@@ -118,7 +118,7 @@ class Cache implements CacheConfigurationInterface
     /**
      * Return's the reference to the configuration instance.
      *
-     * @return \TechDivision\Import\ConfigurationInterface The configuration instance
+     * @return \TechDivision\Import\Configuration\ConfigurationInterface The configuration instance
      */
     public function getConfiguration()
     {