+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add #PAC-85
 * Switch to latest techdivision/import 16.* version as dependency
 * Add new method FileResolver::hasPrefix() to query whethr or not the file resolver has the default prefix
+* Add functionality to also parse configuration files that'll be available in configured additional vendor directory
 
 # Version 15.0.4